Devices/PC/DevQemuFwCfg.cpp: Improve efficiency of the RAM based framebuffer refresh callback by mapping the framebuffer once and copying directly from the memory buffer instead of going through PDMDevHlpPhysReadUser() which requires PGM locking, etc., github:gh-34 [build fix]
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@110323 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/PC/DevQemuFwCfg.cpp
@@ -1757,7 +1757,7 @@ static DECLCALLBACK(int) qemuFwCfgR3RamfbPortUpdateDisplay(PPDMIDISPLAYPORT pInt
             if (RT_UNLIKELY(!pThis->Ramfb.cPgLocks))
             {
                 size_t const cbFb = pThis->RamfbCfg.cbStride * pThis->RamfbCfg.cHeight;
-                uint32_t const cPages = (cbFb + GUEST_PAGE_SIZE - 1) >> GUEST_PAGE_SHIFT;
+                uint32_t const cPages = (uint32_t)((cbFb + GUEST_PAGE_SIZE - 1) >> GUEST_PAGE_SHIFT);
                 RTGCPHYS GCPhysStart = pThis->RamfbCfg.GCPhysRamfbBase & ~(RTGCPHYS)GUEST_PAGE_OFFSET_MASK;
                 RTGCPHYS const offPage = pThis->RamfbCfg.GCPhysRamfbBase & (RTGCPHYS)GUEST_PAGE_OFFSET_MASK;
